got changeStudentGrade to work
